+++ src/view/menu.rs
@@ -106,7 +106,7 @@ impl Menu {
 
         let max_width = 2 * width as i32 / 3;
         let free_width = padding + 2 * border_thickness +
-                         entries.iter().map(|e| font.plan(e.text(), None, None).width as i32)
+                         entries.iter().map(|e| font.plan(e.text(), None, None).width)
                                 .max().unwrap();
 
         let entry_width = free_width.min(max_width);
